[fix][deps] Restored protobuf dependency to 2.4.1 from 2.5.0. More below
Why
===

A cleanup of the `pom.xml` was accompanied by a number of version bumps.
These version bumps did not break the building of chronos, so all was
assumed to be fine. However, upon upgrading we noticed strange and
mysterious breakage related to zookeeper. We determined the protobuf
version bump to be responsible for these breakages and the version has
since been reverted to 2.4.1.