Prevent resetting resultsCollection to save render
Resetting app.resultsCollection forces a full re-render of the job list,
which means two full list renders on start up of the app. Navigating to
the index only needs to reset the detailsCollection and remove
highlights from the list. The detailsCollection was already being reset,
and the highlight is moved in this commit to the reset event in the
JobsCollectionView.