Support easy submission of immediate job from REST API
This is a nice feature for job submission go through the REST API
because if a job want to schedule now, it has first to know what's the
time the chronos server is, and give the job a schedule time within
epsilon window. This is a problem when submitting a job from web ui,
because the schedule time by default is current time. So it's cool if
we can do that through REST API.

With this patch, one job can be a immediate job by any of below ways:
  * without "schedule" field
  * with empty start time in "schedule" field, e.g. "R1//PT24H"

Change-Id: I36e5b2996048b3166c90b240fb2d682da8f34d76
Signed-off-by: Chengwei Yang <yangchengwei@qiyi.com>