Update to the latest version of scala/mesos
We need to update to the latest version of scala, mesos, and other
relevant libraries so we can use the latest features available in mesos
and scala. This includes the latest docker support as well as bridged
networking with docker. Also, with the latest version of scala, this
removes the limitation on the number of fields in a case class.