Set the offer frequency to 5 secs
When an offer is refused, we wait 5 secs before receiving that offer
again (consistent with Marathon). This will have an effect on triggered
jobs, in that they will be a short delay in triggering them. However,
it will ensure that Chronos doesn’t receive and decline offers too
frequently and that Chronos doesn’t receive more than it’s fair share
of offers.