Forbid using canBeReadonly when connecting zookeeper
Chronos may get a wrong leader path, which may cause a fatal error that:
once the non-leader instance gets an out-of-date leader path and the wrong
path happens to be itself, it will redirects a request to itself  and
then blocks other any requests in less than 1 second. In this circumstance,
scheduler is still running but no rest api responds, even it becomes leader.

So, setting connecting parameter 'canBeReadOnly' to false(default) to provide
Chronos from reading wrong leader info is necessary.