fix a formatting issue with api docs
Prior to this commit, there was a Markdown formatting problem with the
API docs. This resulted in a bunch of (what should have been) bullet
points appearing as normal text, and making the explanation of the ISO
8601 format harder to read than it needed to be.

This commit modifies some of the API documentation to correctly format
the bullet list for the "Adding a Scheduled Job" section.